move lxd related files to lxd_cloud dir
