+++ src/lxd_cloud/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2024 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Module for managing local LXD cloud."""
