+++ src/lxd_cloud/shared_fs.py
